+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `opentelemetry-instrumentation-asyncio` Fix duplicate instrumentation
   ([#3383](https://github.com/open-telemetry/opentelemetry-python-contrib/issues/3383))
 
+- `opentelemetry-instrumentation-botocore` Add GenAI instrumentation for additional Bedrock models for InvokeModel API
+  ([#3419](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3419))
+
 ## Version 1.32.0/0.53b0 (2025-04-10)
 
 ### Added